feat(ssr): move monkeypatching inside code
