+++ src/instantsearch.js
@@ -1,3 +1,4 @@
 export { createWidgetMixin } from './mixins/widget';
 export * from './widgets';
 export { plugin as default } from './plugin';
+export { createInstantSearch } from './util/createInstantSearch';
